Editor and FontInfo views share a lot of commonality, extract it
